requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt
pydantic
supabase
pydantic_settings